% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMRF.R
\name{calc_pairwise_distance}
\alias{calc_pairwise_distance}
\title{Calculate pairwise distance matrix for
a given spatial coordinate of transcriptomic data.}
\usage{
calc_pairwise_distance(spatial_coords)
}
\arguments{
\item{spatial_coords}{A matrix or data frame where each row represents a point in space with coordinates.}
}
\value{
A distance matrix where entry (i, j) represents the distance between spatial coordinates i and j.
}
\description{
Calculate pairwise distance matrix for
a given spatial coordinate of transcriptomic data.
}
\examples{
spatial_coords <- matrix(c(1, 2, 3, 4, 5, 6), ncol = 2, byrow = TRUE)
calc_pairwise_distance(spatial_coords)
}
