% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier_celltype_determination.R
\name{celltype_test}
\alias{celltype_test}
\title{Regression Test for Identifying Significant Mutations and Carrier Celltypes.}
\usage{
celltype_test(
  celltypes = NULL,
  voi = NULL,
  N = NULL,
  vaf = NULL,
  X = NULL,
  Ws = NULL,
  spatial_coords = NULL,
  test_type = c("linear", "weighted"),
  permute_num = 1000,
  plot = F,
  save_path = NULL,
  figure_height = 10,
  figure_width = 10,
  method = "Raw"
)
}
\arguments{
\item{celltypes}{A string vector containing all unique celltypes.}

\item{voi}{A string vector containing names of variant of interest. Usually a subset of variants that went through basic filtering.}

\item{N}{A MT variant x cell total coverage matrix.}

\item{vaf}{A MT variant x cell variant allele frequnecy (VAF) matrix. Equivalent as af.dm matrix.}

\item{X}{A MT variant x cell alternative allele count matrix. Optional if vaf matrix is provided.}

\item{Ws}{A cell x celltype weight matrix. Representing spatial decomposition results from tools like RCTD.}

\item{spatial_coords}{A cell x spatial coordinates matrix. Representing the actual spot/cell location in space.}

\item{test_type}{Type of Celltype test. Options from c("linear","weighted").}

\item{permute_num}{A numeric number representing permuation number for each variant.}

\item{save_path}{path for output results}

\item{figure_height}{output pdf figure height}

\item{figure_width}{output pdf figure width}

\item{method}{P-value correction method: Options: c("FDR", "FWER", "Raw").}

\item{...}{Other options used to control matching behavior between duplicate
strings. Passed on to \code{\link[stringi:stri_opts_collator]{stringi::stri_opts_collator()}}.}
}
\value{
A list object containing fitted parameters (intercept, coefficient and p-values).
}
\description{
\code{celltype_test()} implements the statistical tests for identifying significant celltypes given spatial decomposition and varaint information.
}
\examples{
celltype_test(celltypes=c("BE","DYSP","Normal"), voi=voi_list,
N=N,vaf=af.dm, test_type="linear",permute_num=1000)

}
