% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier_celltype_determination.R
\name{celltype_test_knn}
\alias{celltype_test_knn}
\title{Perform Celltype Test Using Nearest Neighbors for Each Spot}
\usage{
celltype_test_knn(
  celltypes,
  vars,
  N_voi,
  vaf,
  Ws,
  spatial_coords_temp,
  test_type = c("linear", "weighted"),
  permute = F,
  permute_num = 1000,
  k_neighbors = 100,
  method = "Raw",
  sample_idx = NULL,
  disease_celltype = "BE",
  ratio_threshold = 0.03,
  vaf_cellprop = F,
  exclude_plot_idx = NULL,
  p_thresh = 0.05,
  max_log10p_cap = 6,
  coef_plot_option = "negative"
)
}
\arguments{
\item{celltypes}{A string vector of unique cell types.}

\item{vaf}{A matrix of variant allele frequencies (MT variant x cell).}

\item{Ws}{A matrix of cell type weights (cell x celltype).}

\item{spatial_coords_temp}{A matrix or data frame of spatial coordinates (cell x 2) with row names as cell IDs.}

\item{test_type}{A string specifying the regression test: "linear" or "weighted".}

\item{permute_num}{Number of permutations for the weighted test (default 1000).}

\item{k_neighbors}{Number of nearest neighbors to use for each spot (default = 100).}

\item{method}{P-value correction method. Options are "FDR", "FWER", or "Raw" (default = "Raw").}

\item{voi}{A string vector of variants of interest.}

\item{N}{A matrix of total coverage counts (MT variant x cell).}
}
\value{
A list with one element per spot (named by the spot ID). Each element is a list containing:
\item{intercept}{A matrix of intercept estimates (rows = variants, columns = cell types).}
\item{coef}{A matrix of regression coefficients.}
\item{pval}{A matrix of raw p-values.}
\item{adjusted_pval}{A matrix of adjusted p-values (if method != "Raw"); otherwise, same as raw.}
}
\description{
This function performs a celltype test for each spot by selecting its nearest k neighbors
(including itself) and running a linear (or weighted) regression of variant allele frequency
versus cell type weights for each variant of interest. The regression is performed for each cell
type. Optionally, p-values across all tests are adjusted using the specified method.
}
\examples{
\dontrun{
  # results_knn <- celltype_test_knn(celltypes, voi, N, vaf, Ws, spatial_coords_temp,
  #                                  test_type = "linear", permute_num = 1000,
  #                                  k_neighbors = 100, method = "FDR")
}

}
