% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier_celltype_determination.R
\name{compute_gene_program_enrichment}
\alias{compute_gene_program_enrichment}
\title{Compute Gene Program Enrichment per Grid}
\usage{
compute_gene_program_enrichment(
  seu,
  results_grid,
  spatial_coords_grids,
  gene_programs
)
}
\arguments{
\item{seu}{Seurat object containing gene expression data.}

\item{results_grid}{A list containing grid-based cell assignments.}

\item{gene_programs}{A named list where each element is a vector of genes for a specific program.}
}
\value{
A list containing enrichment scores for each grid.
}
\description{
Compute Gene Program Enrichment per Grid
}
