% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier_celltype_determination.R
\name{filter_significant_variants}
\alias{filter_significant_variants}
\title{Filter Significant Variants Based on Adjusted P-values and Coefficients}
\usage{
filter_significant_variants(
  results,
  alpha_threshold = 0.05,
  output_file = "output/significant_variants.txt"
)
}
\arguments{
\item{results}{A list object returned by \code{celltype_test()} that contains matrices:
\code{adjusted_pval} (adjusted p-values) and \code{coef} (regression coefficients) with rows as variants
and columns as cell types.}

\item{alpha_threshold}{A numeric value specifying the significance threshold for adjusted p-values (default is 0.05).}

\item{output_file}{A character string specifying the filename where the significant variants will be saved (default is "significant_variants.txt").}
}
\value{
A list where each element corresponds to a cell type and contains the names of the variants that meet the criteria.
}
\description{
This function takes a results object (from a celltype testing analysis) that contains raw and adjusted
p-values as well as coefficients for each variant across multiple cell types. It filters for variants
with an adjusted p-value below a specified alpha threshold and a positive coefficient, and then saves
the significant variants for each cell type into a text file.
}
\examples{
\dontrun{
  # With 'results'object returned from celltype_test():
  sig_vars <- filter_significant_variants(results, alpha_threshold = 0.05, output_file = "output/sig_variants.txt")
}

}
