% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier_celltype_determination.R
\name{plot_pval_grid}
\alias{plot_pval_grid}
\title{Plot P-Values from Grid-Based Regression}
\usage{
plot_pval_grid(results_grid, var_name = NULL, grid_size = 20)
}
\arguments{
\item{results_grid}{A list containing p-values for each spatial grid.}

\item{grid_size}{Number of bins per axis (default = 20 for 20x20 grid).}
}
\value{
A \code{ggplot2} object showing p-value distribution across the spatial grid.
}
\description{
This function plots the mean p-value per grid after running the \code{celltype_test_grid} function.
}
