% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plot.R
\name{plot_vaf_cellprop}
\alias{plot_vaf_cellprop}
\title{Helper functions of the spatial MT model}
\usage{
plot_vaf_cellprop(
  i,
  af.dm,
  norm_weights,
  plot_df,
  intercept,
  coef,
  pval,
  permute = F
)
}
\arguments{
\item{i}{A numeric indicator of variant.}

\item{af.dm}{A MT variant x cell variant allele frequnecy (VAF) matrix. Equivalent as vaf matrix.}

\item{norm_weights}{A cell x celltype weight matrix, representing spatial decomposition results from tools like RCTD.}

\item{plot_df}{A plotting dataframe.}

\item{voi}{A string vector containing names of variant of interest. Usually a subset of variants that went through basic filtering.}
}
\description{
\code{plot_vaf_cellprop()} plots the VAF vs coverage for a given celltype and given variant.
}
