% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineagePlot.R
\name{PlotBayesFactor}
\alias{PlotBayesFactor}
\title{Plot Simple Model Bayes Factor}
\usage{
PlotBayesFactor(var, N, af.dm, spatial_coords, BF_cap = 3, plot = F)
}
\arguments{
\item{var}{Character string, the variant of interest.}

\item{N}{Matrix or data frame, the total counts for each spot.}

\item{af.dm}{Matrix or data frame, the allele frequency data.}

\item{spatial_coords}{Data frame or matrix with spatial coordinates (SPATIAL_1, SPATIAL_2).}

\item{BF_cap}{Numeric, maximum value to cap the Bayes Factor at (default: 3).}
}
\value{
A \code{ggplot2} object showing the spatial distribution of the Bayes Factor.
}
\description{
This function computes and visualizes the Bayes Factor (BF) for a given variant
in spatial mitochondtial data. The BF is calculated using a binomial
likelihood ratio test, comparing a baseline probability (\code{p0}) estimated from
all cells to the maximum likelihood estimate (\code{p1}) for each spatial spot.
}
\examples{
PlotBayesFactor("3054_G>C", N, af.dm, spatial_coords.plot=T)
}
