% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier_celltype_determination.R
\name{power_analysis_all}
\alias{power_analysis_all}
\title{Power Analysis for Identifying Negative Mutations and Carrier Celltypes.
for all variants of interest & all celltypes
set permute_num=20 for now, shall be 1000}
\usage{
power_analysis_all(
  voi = NULL,
  celltypes = NULL,
  Ws = NULL,
  N = NULL,
  vaf = NULL,
  X = NULL,
  sample_num = 100,
  alpha = 0.05,
  n_sim = 10,
  beta_threshold = 0.5,
  plot = T,
  save_path = NULL,
  height = 8,
  width = 8
)
}
\description{
Power Analysis for Identifying Negative Mutations and Carrier Celltypes.
for all variants of interest & all celltypes
set permute_num=20 for now, shall be 1000
}
