% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutations.R
\name{UpdatePath}
\alias{UpdatePath}
\title{Compute a Simple Path for the spots Using MST and DFS.}
\usage{
UpdatePath(spatial_coords_temp, common_cells, plot = T)
}
\arguments{
\item{plot}{Logical (default: TRUE). If TRUE, plots the path over the spatial coordinates.}

\item{spatial_coords}{A matrix or data frame with two columns representing
the spatial coordinates (e.g., X and Y positions).}
}
\value{
A vector of indices representing the order in which the coordinates are visited.
}
\description{
This function computes a traversal path through spatial coordinates using a
Minimum Spanning Tree (MST) and Depth-First Search (DFS). The path is not
necessarily the shortest but provides a simple and efficient ordering.
}
\examples{
spatial_coords <- matrix(runif(20), ncol=2)
path_order <- mst_dfs_path(spatial_coords, plot=TRUE)

}
