% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier_celltype_determination.R
\name{calc_power}
\alias{calc_power}
\title{Power Analysis for Identifying Negative Mutations and Carrier Celltypes.}
\usage{
calc_power(beta, Nj, Wj, vaf_j, alpha, n_sim, verbose = T, report_frac = 3)
}
\description{
Power Analysis for Identifying Negative Mutations and Carrier Celltypes.
}
