% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier_celltype_determination.R
\name{plot_pval_grid_per_celltype}
\alias{plot_pval_grid_per_celltype}
\title{Plot P-Values from Grid-Based Regression for Each Cell Type}
\usage{
plot_pval_grid_per_celltype(results_grid, grid_size = 20)
}
\arguments{
\item{results_grid}{A list containing p-values for each spatial grid.}

\item{grid_size}{Number of bins per axis (default = 20 for 20x20 grid).}
}
\value{
A \code{ggplot2} object showing p-value distributions for each cell type.
}
\description{
This function plots the -log10 p-values for each cell type and variant
within the spatial grid, using a continuous color scale.
}
