% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plot.R
\name{plot_vaf_cellprop}
\alias{plot_vaf_cellprop}
\title{helper functions of the spatial mito model
plotting VAF vs celltype proportions, colored by coverage}
\usage{
plot_vaf_cellprop(
  i,
  af.dm,
  norm_weights,
  plot_df,
  intercept,
  coef,
  pval,
  permute = F,
  return_plot = F
)
}
\description{
helper functions of the spatial mito model
plotting VAF vs celltype proportions, colored by coverage
}
